+++ src/main/java/orm/dsl/ddl/CreateTableImpl.java
@@ -5,7 +5,7 @@ import orm.TableEntity;
 import orm.TableField;
 import orm.TablePrimaryField;
 import orm.dsl.QueryBuilder;
-import orm.dsl.dialect.h2.ColumnTypeMapper;
+import orm.dsl.sql_dialect.h2.ColumnTypeMapper;
 import orm.dsl.step.ddl.CreateTableStep;
 import orm.exception.InvalidIdGenerationException;
